Moved tests to bnd
