--- addons/binding/org.openhab.binding.tradfri.test/META-INF/MANIFEST.MF
@@ -1,26 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: TRÅDFRI Binding Tests
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.openhab.binding.tradfri.test;singleto
- n:=true
-Bundle-Vendor: openHAB
-Bundle-Version: 2.5.0.qualifier
-Export-Package: org.openhab.binding.tradfri;uses:="org.eclipse.smarthome.test"
-Fragment-Host: org.openhab.binding.tradfri
-Import-Package: 
- org.eclipse.jdt.annotation;resolution:=optional,
- org.eclipse.smarthome.core.common.registry,
- org.eclipse.smarthome.core.events,
- org.eclipse.smarthome.core.thing.util,
- org.eclipse.smarthome.test.java,
- org.eclipse.smarthome.test.storage,
- org.hamcrest;core=split,
- org.hamcrest.core,
- org.junit;version="4.0.0",
- org.mockito,
- org.mockito.stubbing,
- org.osgi.framework,
- org.osgi.service.device,
- org.slf4j
-Automatic-Module-Name: org.openhab.binding.tradfri.test
