--- addons/binding/org.openhab.binding.tradfri.test/build.properties
@@ -1,5 +1,0 @@
-source.. = src/test/java/
-output.. = target/classes
-bin.includes = META-INF/,\
-               .,\
-               NOTICE
