+++ itests/org.openhab.binding.tradfri.tests/NOTICE
